refactor: Station 패키지 변경
