+++ src/main/java/subway/line/LineService.java
@@ -5,9 +5,9 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.Station;
-import subway.StationRepository;
-import subway.StationResponse;
+import subway.station.Station;
+import subway.station.StationRepository;
+import subway.station.StationResponse;
 
 @Service
 @RequiredArgsConstructor
