+++ src/main/java/subway/station/Station.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
