+++ src/main/java/subway/station/StationRepository.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
