+++ src/main/java/subway/station/SubwayApplication.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
